requirements.txt
openai-whisper==20230117
requests==2.28.2
fuzzysearch==0.7.3
